Config table view cell.
